setup basic structure for WatchGame and JoinGame in GameView
