+++ frontend/src/components/WatchGame.vue
@@ -1,0 +1,2 @@
+<template>This is Watch Game</template>
+<script setup lang="ts"></script>
